Rename unlock time to unlock height, and make it uint32_t (it's height only, not timestamp)

+++ include/ITransaction.h
@@ -63,7 +63,7 @@ public:
   virtual Crypto::Hash getTransactionInputsHash() const = 0;
   virtual Crypto::PublicKey getTransactionPublicKey() const = 0;
   virtual bool getTransactionSecretKey(Crypto::SecretKey& key) const = 0;
-  virtual uint64_t getUnlockTime() const = 0;
+  virtual uint32_t getUnlockTime() const = 0;
 
   // extra
   virtual bool getPaymentId(Crypto::Hash& paymentId) const = 0;
@@ -121,10 +121,10 @@ public:
   virtual size_t addInput(const MultisignatureInput& input) = 0;
   virtual size_t addInput(const AccountKeys& senderKeys, const TransactionTypes::InputKeyInfo& info, KeyPair& ephKeys) = 0;
 
-  virtual size_t addOutput(uint64_t amount, const AccountPublicAddress& to, uint64_t unlockTime) = 0;
-  virtual size_t addOutput(uint64_t amount, const std::vector<AccountPublicAddress>& to, uint32_t requiredSignatures, uint64_t unlockTime) = 0;
-  virtual size_t addOutput(uint64_t amount, const KeyOutput& out, uint64_t unlockTime) = 0;
-  virtual size_t addOutput(uint64_t amount, const MultisignatureOutput& out, uint64_t unlockTime) = 0;
+  virtual size_t addOutput(uint64_t amount, const AccountPublicAddress& to, uint32_t unlockTime) = 0;
+  virtual size_t addOutput(uint64_t amount, const std::vector<AccountPublicAddress>& to, uint32_t requiredSignatures, uint32_t unlockTime) = 0;
+  virtual size_t addOutput(uint64_t amount, const KeyOutput& out, uint32_t unlockTime) = 0;
+  virtual size_t addOutput(uint64_t amount, const MultisignatureOutput& out, uint32_t unlockTime) = 0;
 
   // transaction info
   virtual void setTransactionSecretKey(const Crypto::SecretKey& key) = 0;
